Introduce model traits
