+++ app/Model/Trait/Draftable.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Model\Trait;
+ 
+trait Draftable {
+
+    public function isDraft(){
+        return $this->active == 0;
+    }
+
+}
