+++ composer.json
@@ -19,7 +19,8 @@
         "nesbot/carbon": "2.*",
         "mews/purifier": "^3.3",
         "laravel/ui": "^3.0",
-        "codedge/laravel-selfupdater": "^3.4"
+        "codedge/laravel-selfupdater": "^3.4",
+        "vyuldashev/laravel-openapi": "1.12.0"
     },
     "require-dev": {
         "phpunit/phpunit" : "^9.0",
